Adjust lat & lon data type to int

+++ src/main/java/org/littil/api/location/repository/AbstractLocationEntity.java
@@ -17,8 +17,8 @@ import javax.persistence.MappedSuperclass;
 public abstract class AbstractLocationEntity extends AbstractAuditableEntity {
 
     @Column(name = "latitude")
-    private Double latitude;
+    private int latitude;
 
     @Column(name = "longitude")
-    private Double longitude;
+    private int longitude;
 }
